requirements.txt
Package             Version
------------------- -------
asgiref             3.4.1
cycler              0.11.0
Django              3.2.10
django-cors-headers 3.10.1
django-filter       21.1
djangorestframework 3.12.4
Faker               11.3.0
fonttools           4.28.5
kiwisolver          1.3.2
matplotlib          3.5.1
mixer               7.2.0
null                0.6.1
numpy               1.22.0
packaging           21.3
password            0.2
Pillow              9.0.0
pip                 21.3.1
pyparsing           3.0.6
python-dateutil     2.8.2
pytz                2021.3
response            0.5.0
scipy               1.7.3
setuptools          59.5.0
six                 1.16.0
sqlparse            0.4.2
text-unidecode      1.3
username            2.1.0
wheel               0.37.0
